Create post-sort-on-acf
Add this to function.php in order to use Advanced Custom Fields in Elementor Queries. 

Be sure to have a custom date field created. In our example, we use, “sort_date”.
Go to the page/template in Elementor that you want your posts widget on. Give a name in the “Query ID” field. In our example, we use, “sorted_by_date”.
Lastly, go to your theme’s functions.php file and add the code.